requirements.txt
certifi==2019.11.28
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
numpy==1.18.2
openpyxl==3.0.3
pandas==1.0.3
pyaml==20.3.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.12.0
tabulate==0.8.7
urllib3==1.25.8
